updating to newest version

+++ Pirates/MovableObject.cpp
@@ -1,6 +1,6 @@
 #include "MovableObject.h"
 
-MovableObject::MovableObject(sf::Texture *texture, int speed, int healthMax, int healthCurrent)
+MovableObject::MovableObject(sf::Texture *texture, float speed, int healthMax, int healthCurrent)
 	: Object(texture), healthCurrent(healthCurrent), healthMax(healthMax), speed(speed)
 {
 	// nothing
